+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Releases
 
 - **Added:** Rule `DOC012` to check for leading in first non-blank line of docstring.
+- **Added:** Rule `DOC106` to check parameter order mismatch with function signature.
 
 ### v0.0.17 (2025-11-07)